Changed background color for preview rendering

+++ ICE/src/Graphics/API/GraphicsAPI.h
@@ -20,7 +20,7 @@ namespace ICE {
         virtual void initialize() const = 0;
         virtual void bindDefaultFramebuffer() const = 0;
         virtual void setViewport(int x, int y, int width, int height) const = 0;
-        virtual void setClearColor(int r, int g, int b, int a) const = 0;
+        virtual void setClearColor(float r, float g, float b, float a) const = 0;
         virtual void clear() const = 0;
         virtual void renderVertexArray(const VertexArray* va) const = 0;
         virtual void flush() const = 0;
